Fix struct alignment in win_perf_counters for 32-bit 386 arch (#4206)
